+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6 (Pending):
+
+## 0.5:
+* Add SessionManager.getSessionIDs() to expose the list of sessions in manager.
+* Add SessionInfo.getAttributeKeys() to expose the list of attributes in session.
+
 ## 0.4:
 * Avoid marking InMemorySessionManager as Serializable as not all sub-classes are Serializable.